Reverted all changes in animation_key_frames and moved new code to a new module in dedicated package.
