+++ scripts/deforum_helpers/render_tools/__init__.py
@@ -1,3 +1,4 @@
+from .animation_keys import AnimationKeys
 from .animation_mode import AnimationMode
 from .schedule import Schedule
 from .srt import Srt
