+++ scripts/deforum_helpers/render_tools/animation_keys.py
@@ -1,0 +1,33 @@
+import dataclasses
+
+from ..animation_key_frames import DeformAnimKeys, LooperAnimKeys
+
+
+@dataclasses.dataclass
+class AnimationKeys:
+    def __init__(self, deform_keys: DeformAnimKeys, looper_keys: LooperAnimKeys):
+        self.deform_keys = deform_keys  # Not a typo. It's about deforming the frame.
+        self.looper_keys = looper_keys
+
+    def update(self, i: int):
+        self.looper_keys.use_looper = self.looper_keys.use_looper
+        self.looper_keys.imagesToKeyframe = self.looper_keys.imagesToKeyframe
+        # TODO FIXME refactor index handling and remove new boilerplate
+        self.looper_keys.imageStrength = self.looper_keys.image_strength_schedule_series[i]
+        self.looper_keys.blendFactorMax = self.looper_keys.blendFactorMax_series[i]
+        self.looper_keys.blendFactorSlope = self.looper_keys.blendFactorSlope_series[i]
+        self.looper_keys.tweeningFramesSchedule = self.looper_keys.tweening_frames_schedule_series[i]
+        self.looper_keys.colorCorrectionFactor = self.looper_keys.color_correction_factor_series[i]
+
+    @staticmethod
+    def choose_default_or_parseq_keys(default_keys, parseq_keys, parseq_adapter):
+        return default_keys if not parseq_adapter.use_parseq else parseq_keys
+
+    @staticmethod
+    def from_args(anim_args, loop_args, parseq_adapter, seed):
+        # Parseq keys are decorated, see ParseqAinmKeysDecorator and ParseqLooperKeysDecorator
+        return AnimationKeys(
+            AnimationKeys.choose_default_or_parseq_keys(DeformAnimKeys(anim_args, seed),
+                                                        parseq_adapter.anim_keys, parseq_adapter),
+            AnimationKeys.choose_default_or_parseq_keys(LooperAnimKeys(loop_args, anim_args, seed),
+                                                        parseq_adapter.looper_keys, parseq_adapter))
